Fold in version/changelog validation
